Initial commit
A lot of stubs for the GUI, just enough to have a '+' button, adding a
new podcast using Tracker.
The miner is almost done, and is greatly inspired by the RSS tracker.

The icon is a variation of the Music application icon, with color and
elements reminding of RSS.

The design of the GUI is inspired by both the Music and News
applications.